8886: Experimental asynchronous permissions update.
Add configuration parameter 'async_permissions_update' (default false).  If
true, do not delete permission cache in #invalidate_permissions_cache, but
instead trigger "NOTIFY invalidate_permissions_cache" on the database.

Add script/permission-updater.rb which runs as an independent process.  It
blocks on "LISTEN invalidate_permissions_cache" and updates the permission
cache whenever notified.

This is not ready for use; in particular it creates a race condition
recomputing permissions with effects such as not being able to read back API
records that were just created.

+++ services/api/app/models/user.rb
@@ -124,18 +124,14 @@ class User < ArvadosModel
   end
 
   def self.invalidate_permissions_cache
-    Rails.cache.delete_matched(/^groups_for_user_/)
+    if Rails.configuration.async_permissions_update
+      connection.execute "NOTIFY invalidate_permissions_cache"
+    else
+      Rails.cache.delete_matched(/^groups_for_user_/)
+    end
   end
 
-  # Return a hash of {group_uuid: perm_hash} where perm_hash[:read]
-  # and perm_hash[:write] are true if this user can read and write
-  # objects owned by group_uuid.
-  #
-  # The permission graph is built by repeatedly enumerating all
-  # permission links reachable from self.uuid, and then calling
-  # search_permissions
-  def group_permissions
-    Rails.cache.fetch "groups_for_user_#{self.uuid}" do
+  def calculate_group_permissions
       permissions_from = {}
       todo = {self.uuid => true}
       done = {}
@@ -182,7 +178,21 @@ class User < ArvadosModel
           end
         end
       end
-      search_permissions(self.uuid, permissions_from)
+      perms = search_permissions(self.uuid, permissions_from)
+      Rails.cache.write "groups_for_user_#{self.uuid}", perms
+      perms
+  end
+
+  # Return a hash of {group_uuid: perm_hash} where perm_hash[:read]
+  # and perm_hash[:write] are true if this user can read and write
+  # objects owned by group_uuid.
+  #
+  # The permission graph is built by repeatedly enumerating all
+  # permission links reachable from self.uuid, and then calling
+  # search_permissions
+  def group_permissions
+    Rails.cache.fetch "groups_for_user_#{self.uuid}" do
+      calculate_group_permissions
     end
   end
